requirements.txt
discord.py[voice]==1.6.0
